requirements.txt
numpy>=1.24
scipy>=1.10
numba>=0.57
moderngl>=5.8
dearpygui>=1.9
colour-science>=0.4
imageio>=2.28
Pillow>=9.5
scikit-image>=0.21
matplotlib>=3.7
pyglm>=2.7
